Move notification_history method into dao file for consistency
